Rename to packet_helpers
