+++ crates/nym-connection-monitor/src/icmp_beacon.rs
@@ -19,7 +19,7 @@ use tracing::{debug, error, trace};
 
 use crate::{
     error::Result,
-    packet::{
+    packet_helpers::{
         create_icmpv4_echo_request, create_icmpv6_echo_request, is_icmp_echo_reply,
         is_icmp_v6_echo_reply, wrap_icmp_in_ipv4, wrap_icmp_in_ipv6,
     },
